+++ sdk/src/main/java/io/stipop/adapter/StickerThumbAdapter.kt
@@ -3,7 +3,6 @@ package io.stipop.adapter
 import android.annotation.SuppressLint
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import io.stipop.PackUtils
 import io.stipop.models.SPSticker
 import io.stipop.models.Sticker
 import io.stipop.models.StickerPackage
