+++ sdk/src/main/java/io/stipop/view/viewmodel/PackageDetailViewModel.kt
@@ -3,7 +3,7 @@ package io.stipop.view.viewmodel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import io.stipop.PackUtils
+import io.stipop.StipopUtils
 import io.stipop.Stipop
 import io.stipop.api.StipopApi
 import io.stipop.data.StickerDetailRepository
@@ -42,7 +42,7 @@ internal class PackageDetailViewModel(private val repository: StickerDetailRepos
             stickerPackage.value?.let { stickerPackage ->
                 if (!stickerPackage.isDownloaded()) {
                     repository.postDownloadStickers(stickerPackage) {
-                        PackUtils.downloadAndSaveLocalV2(stickerPackage) {
+                        StipopUtils.downloadAtLocal(stickerPackage) {
                             PackageDownloadEvent.publishEvent(stickerPackage.packageId)
                         }
                     }
