+++ sdk/src/main/res/layout/view_keyboard_popup.xml
@@ -18,6 +18,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
+
         <RelativeLayout
             android:id="@+id/recent_favorite_container"
             android:layout_width="48dp"
